ltt newlib commit it, it compiles, not tested
git-svn-id: http://ltt.polymtl.ca/svn@1011 04897980-b3bd-0310-b5e0-8ef037075253

+++ ltt/branches/poly/ltt/event.h
@@ -19,6 +19,7 @@
 #ifndef EVENT_H
 #define EVENT_H
 
+#include <glib.h>
 #include <ltt/ltt.h>
 
 LttEvent *ltt_event_new();
@@ -96,7 +97,7 @@ void *ltt_event_data(LttEvent *e);
    instance. This function returns the number of elements for an array or 
    sequence field in an event. */
 
-unsigned ltt_event_field_element_number(LttEvent *e, LttField *f);
+guint64 ltt_event_field_element_number(LttEvent *e, LttField *f);
 
 
 /* Set the currently selected element for a sequence or array field. */
